refactor: 457 - move all data files to dev-data folder
