+++ dev-data/about.data.ts
@@ -1,11 +1,11 @@
+import type { CourseNamesChannels } from './communication.data';
+import { DISCORD_LINKS } from './communication.data';
 import awardIcon from '@/shared/assets/icons/award-icon.webp';
 import giftIcon from '@/shared/assets/icons/gift.webp';
 import noteIcon from '@/shared/assets/icons/note-icon.webp';
 import paperIcon from '@/shared/assets/icons/paper-icon.webp';
 import personIcon from '@/shared/assets/icons/person-icon.webp';
 import planetIcon from '@/shared/assets/icons/planet.webp';
-import type { CourseNames } from '@/shared/data/communication.data';
-import { DISCORD_LINKS } from '@/shared/data/communication.data';
 
 type AboutInfo = {
   id: number;
@@ -15,7 +15,7 @@ type AboutInfo = {
 };
 
 type ContentMap = {
-  [key in CourseNames]: AboutInfo[];
+  [key in CourseNamesChannels]: AboutInfo[];
 };
 
 const angularNodejsAwsFundamentals: (course: string) => AboutInfo[] = () => [
@@ -196,7 +196,7 @@ const reactRuAbout: AboutInfo[] = [
   },
 ];
 
-export const contentMap: ContentMap = {
+export const contentMapAbout: ContentMap = {
   'js / front-end ru': javaScriptRU(),
   'js / front-end en': javaScriptEN(),
   'js / front-end pre-school ru': javaScriptPreSchoolRU(),
