+++ dev-data/aws-dev-path.data.ts
