+++ dev-data/courses-data.types.ts
