+++ dev-data/courses-path.data.ts
