+++ dev-data/index.ts
@@ -1,0 +1,22 @@
+export { type AngularAwsPath, type CoursesPath, type DataMap, type JSPath } from './courses-data.types';
+export { COURSE_TITLES } from './courseTitles.data';
+export { type CourseNames, contentMap } from './training-program.data';
+export { type CourseNamesChannels, DISCORD_LINKS, RS_DOCS_COMMUNICATION_LINK, RS_DOCS_TELEGRAM_CHATS_LINK } from './communication.data';
+export { angular } from './angular.data';
+export { angularPath } from './angular-path.data';
+export { awsDev } from './awsDev.data';
+export { awsDevPath } from './aws-dev-path.data';
+export { awsFundamentals } from './awsFundamentals.data';
+export { contentMapAbout } from './about.data';
+export { courseDataMap } from './required.data';
+export { courses } from './courses.data';
+export { coursesPath } from './courses-path.data';
+export { events } from './events.data';
+export { faqData } from './faq.data';
+export { javaScriptRu } from './javascript-ru.data';
+export { jsPath } from './js-path.data';
+export { jsPathRu } from './js-ru-path.data';
+export { nodejs } from './nodejs.data';
+export { preSchoolEn, preSchoolRu } from './preSchool.data';
+export { reactEn } from './react-en.data';
+export { reactRu } from './react-ru.data';
