+++ dev-data/js-ru-path.data.ts
