+++ dev-data/preSchool.data.ts
@@ -1,4 +1,4 @@
-import type { Trainer } from './trainers.types';
+import type { Trainer } from '../src/widgets/trainers/trainers.types';
 import preSchoolImg2 from '@/shared/assets/mentors/a-musikhina.webp';
 import preSchoolImg1 from '@/shared/assets/mentors/v-kavaliou.webp';
 import preSchoolImg3 from '@/shared/assets/mentors/v-rogozin.webp';
