--- src/app/services/data/index.ts
@@ -1,7 +1,0 @@
-export { type DataMap } from './courses-data.types';
-export { angularPath } from './angular-path.data';
-export { awsDevPath } from './aws-dev-path.data';
-export { courses } from './courses.data';
-export { coursesPath } from './courses-path.data';
-export { jsPath } from './js-path.data';
-export { jsPathRu } from './js-ru-path.data';
