+++ src/pages/aws-fundamentals.tsx
@@ -5,8 +5,8 @@ import { Communication } from '@/widgets/communication';
 import { CourseMain } from '@/widgets/course-main';
 import { Required } from '@/widgets/required';
 import { Trainers } from '@/widgets/trainers';
-import { awsFundamentals } from '@/widgets/trainers/awsFundamentals.data';
 import { TrainingProgram } from '@/widgets/training-program';
+import { awsFundamentals } from 'data';
 
 const COURSE_NAME = 'aws fundamentals';
