+++ src/shared/hooks/use-data-by-name/use-data-by-name.test.ts
@@ -3,8 +3,8 @@ import { MockedFunction, describe, expect, it, vi } from 'vitest';
 import { useDataByName } from './use-data-by-name';
 import { ROUTES } from '@/app/const';
 import { fetchDataByName } from '@/app/services/api';
-import { DataMap } from '@/app/services/data/courses-data.types';
 import { MOCKED_IMAGE_PATH } from '@/shared/__tests__/constants';
+import { DataMap } from 'data';
 
 const courses: DataMap['courses'] = [
   {
