+++ src/widgets/about/ui/about.tsx
@@ -1,12 +1,12 @@
 /* eslint-disable @stylistic/jsx-one-expression-per-line */
 import { InfoGrid } from './info-grid/info-grid';
-import { contentMap } from '../about.data';
 import { type Course } from '@/app/types';
 import { useCourseByTitle } from '@/shared/hooks/use-course-by-title';
 import { ArrowIcon } from '@/shared/icons';
 import { LinkCustom } from '@/shared/ui/link-custom';
 import { Paragraph } from '@/shared/ui/paragraph';
 import { WidgetTitle } from '@/shared/ui/widget-title';
+import { contentMapAbout } from 'data';
 
 import './about.scss';
 
@@ -51,7 +51,7 @@ export const About = ({ courseName, type = 'en' }: AboutProps) => {
 
   const course = data as Course;
 
-  const infoList = contentMap[courseName];
+  const infoList = contentMapAbout[courseName];
 
   if (loading) {
     return <p>Loading...</p>;
