+++ src/widgets/communication/ui/communication.tsx
@@ -1,21 +1,21 @@
 /* eslint-disable @stylistic/jsx-one-expression-per-line */
-import {
-  CourseNames,
-  DISCORD_LINKS,
-  RS_DOCS_COMMUNICATION_LINK,
-  RS_DOCS_TELEGRAM_CHATS_LINK,
-} from '@/shared/data/communication.data';
 import { TextLinkIcon } from '@/shared/icons';
 import { DiscordLogo } from '@/shared/icons/discord-logo';
 import { LinkCustom } from '@/shared/ui/link-custom';
 import { Paragraph } from '@/shared/ui/paragraph';
 import { Subtitle } from '@/shared/ui/subtitle';
 import { WidgetTitle } from '@/shared/ui/widget-title';
+import {
+  CourseNamesChannels,
+  DISCORD_LINKS,
+  RS_DOCS_COMMUNICATION_LINK,
+  RS_DOCS_TELEGRAM_CHATS_LINK,
+} from 'data';
 
 import './communication.scss';
 
 interface RequiredProps {
-  courseName: CourseNames;
+  courseName: CourseNamesChannels;
   lang?: 'ru' | 'en';
 }
