+++ src/widgets/course-main/courseLoader.ts
@@ -1,4 +1,4 @@
-import { courses } from '@/app/services/data';
+import { courses } from 'data';
 
 export const courseLoader = async () => {
   return courses;
