+++ src/widgets/events/ui/events.tsx
@@ -1,7 +1,6 @@
 import cn from 'classnames';
 import classNames from 'classnames/bind';
 import dayjs from 'dayjs';
-import { events } from '../events.data';
 import { EventCard, EventCardProps } from '@/entities/events';
 import photo3 from '@/shared/assets/photo-3.webp';
 import { getActualData } from '@/shared/helpers/getActualData';
@@ -10,6 +9,7 @@ import { Paragraph } from '@/shared/ui/paragraph';
 import { SectionLabel } from '@/shared/ui/section-label';
 import { Subtitle } from '@/shared/ui/subtitle';
 import { WidgetTitle } from '@/shared/ui/widget-title';
+import { events } from 'data';
 
 import styles from './events.module.scss';
