+++ src/widgets/faq/ui/faq.tsx
@@ -1,8 +1,8 @@
 /* eslint-disable @stylistic/jsx-one-expression-per-line */
 import classNames from 'classnames/bind';
-import { faqData } from '../faq.data';
 import { TextWithLink } from '@/shared/ui/text-with-link';
 import { WidgetTitle } from '@/shared/ui/widget-title';
+import { faqData } from 'data';
 
 import styles from './faq.module.scss';
