+++ src/widgets/training-program/ui/training-program.tsx
@@ -1,11 +1,11 @@
 import { cloneElement } from 'react';
-import { type CourseNames, contentMap } from '../training-program.data';
 import type { Course } from '@/app/types';
 import { useCourseByTitle } from '@/shared/hooks/use-course-by-title';
 import { ArrowIcon } from '@/shared/icons';
 import { Image } from '@/shared/ui/image';
 import { LinkCustom } from '@/shared/ui/link-custom';
 import { WidgetTitle } from '@/shared/ui/widget-title';
+import { type CourseNames, contentMap } from 'data';
 
 import './training-program.scss';
