+++ tsconfig.json
@@ -30,7 +30,8 @@
     "paths": {
       "@/*": [
         "*"
-      ]
+      ],
+      "data": [ "../dev-data" ]
     }
   },
   "include": [
@@ -40,7 +41,9 @@
     "src/shared/__tests__/setup-tests.ts",
     "src/shared/__tests__/visualTesting/main.spec.ts",
     "src/shared/__tests__/visualTesting/navbar.spec.ts",
-    "src/shared/__tests__/visualTesting/footer.spec.ts"
+    "src/shared/__tests__/visualTesting/footer.spec.ts",
+    "dev-data/**/*.ts",
+    "dev-data/**/*.tsx"
   ],
   "exclude": [
     "node_modules"
