+++ vite.config.ts
@@ -15,6 +15,7 @@ export default defineConfig({
   resolve: {
     alias: {
       '@': path.resolve(__dirname, 'src'),
+      'data': path.resolve(__dirname, 'dev-data'),
     },
   },
   css: {
